Update entry scripts and build materials
