+++ .gitignore
@@ -2,7 +2,7 @@ configs/runtime
 **/__pycache__
 ocf_devices/.env
 ocf_devices/dist/*
-ocf_devices/so_lightswitch.egg-info
+ocf_devices/so_demo.egg-info
 ocf_devices/lightswitch/lightswitch_creds
 ocf_devices/**/*.so
 ocf_devices/**/*.o
