+++ ocf_devices/MANIFEST.in
@@ -1,2 +1,2 @@
-graft slined_onboarding/gui/images
+graft slined_onboarding/common/images
 graft slined_onboarding/resources
