+++ ocf_devices/setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = so_lightswitch
+name = so_demo
 version = 0.0.2
 
 [options]
@@ -15,5 +15,5 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-  so_lightswitch_gui = slined_onboarding.gui:gui_main
-# so_lightswitch_cli = slined_onboarding:cli_main
+  so_lightswitch_gui = slined_onboarding.lightswitch:switch_main
+  so_lamp_gui = slined_onboarding.lamp:lamp_main
