+++ ocf_devices/slined_onboarding/lamp/__init__.py
@@ -1,2 +1,3 @@
 from slined_onboarding.lamp.solamp import SoLamp
 from slined_onboarding.lamp.lamp_gui import LampUi
+from slined_onboarding.lamp.lamp_main import lamp_main
